Switch from Lotte to WebDriverJS

+++ jakelib/test.jake
@@ -9,24 +9,26 @@ namespace('test', function() {
   desc('Run all tests');
   task('all', ['bower:install', 'clean:web'], function() {
     return new Promise(function(resolve) {
-      jake.Task['test:white'].addListener('complete', function() {
-        jake.Task['test:black'].addListener('complete', resolve).execute();
+      console.log('Code testing\n------------');
+      jake.Task['test:code'].addListener('complete', function() {
+        console.log('Site testing\n------------')
+        jake.Task['test:site'].addListener('complete', resolve).execute();
       }).execute();
     });
   });
 
-  desc('Run whitebox tests using mocha-phantomjs');
-  task('white', ['bower:install', 'clean:web'], function() {
-    // Start local server
-    var server = spawn('./node_modules/.bin/brunch', 'w -s -e web:dev'.split(' '), {stdio: 'inherit'});
+  desc('Run code-based tests using Mocha PhantomJS');
+  task('code', ['bower:install', 'clean:web'], function() {
+    var server;
 
     return new Promise(function(resolve, reject) {
+      server = spawn('./node_modules/.bin/brunch', 'w -s -e web:dev'.split(' '), {stdio: 'inherit'});
       var testPath = path.resolve(config.overrides['web:dev'].paths.public, 'index.html');
 
       // Catch for Ctrl-C
       process.on('SIGINT', reject);
 
-      // Keep checking to see if tests file are generated.
+      // Keep checking to see if files are generated.
       // Once they are, run mocha-phantomjs.
       var id = setInterval(function() {
         if(fs.existsSync(testPath)) {
@@ -37,41 +39,56 @@ namespace('test', function() {
             command += ' -R ' + process.env.reporter;
           }
 
-          var phantom = jake.createExec(command, {interactive: true}, resolve);
-          phantom.addListener('error', reject);
-          phantom.run();
+          var tester = jake.createExec(command, {interactive: true}, resolve);
+          tester.addListener('error', reject);
+          tester.run();
         }
       }, 1000);
     })
     .finally(function() {
-      server.kill();
+      if(server) {
+        server.kill();
+      }
     });
   });
 
-  desc('Run blackbox tests using Lotte');
-  task('black', function() {
-    // Start local server
-    var server = spawn('./node_modules/.bin/brunch', 'w -s -e web:dev'.split(' '), {stdio: 'inherit'});
+  desc('Run site-based tests using PhantomJS and WebDriverJS');
+  task('site', function() {
+    var phantom;
+    var server;
 
     return new Promise(function(resolve, reject) {
+      phantom = spawn('./node_modules/.bin/phantomjs', ['--webdriver=4444']);
+      server = spawn('./node_modules/.bin/brunch', 'w -s -e web:dev'.split(' '), {stdio: 'inherit'});
       var testPath = path.resolve(config.overrides['web:dev'].paths.public, 'index.html');
 
       // Catch for Ctrl-C
       process.on('SIGINT', reject);
 
-      // Keep checking to see if tests file are generated.
-      // Once they are, run mocha-phantomjs.
+      // Keep checking to see if files are generated.
+      // Once they are, run tests.
       var id = setInterval(function() {
         if(fs.existsSync(testPath)) {
           clearInterval(id);
-          var lotte = jake.createExec('./node_modules/.bin/lotte test/black -I "**/*-test.{coffee,js}"', {interactive: true}, resolve);
-          lotte.addListener('error', reject);
-          lotte.run();
+
+          var command = './node_modules/.bin/mocha';
+          if(process.env.reporter) {
+            command += ' -R ' + process.env.reporter;
+          }
+
+          var tester = jake.createExec(command, {interactive: true}, resolve);
+          tester.addListener('error', reject);
+          tester.run();
         }
       }, 1000);
     })
     .finally(function() {
-      server.kill();
+      if(phantom) {
+        phantom.kill();
+      }
+      if(server) {
+        server.kill();
+      }
     });
   });
 });
